+++ mycli/main.py
@@ -1089,6 +1089,8 @@ class MyCli(object):
               help='Execute command and quit.')
 @click.option('--init-command', type=str,
               help='SQL statement to execute after connecting.')
+@click.option('--charset', type=str,
+              help='Character set for MySQL session.')
 @click.argument('database', default='', nargs=1)
 def cli(database, user, host, port, socket, password, dbname,
         version, verbose, prompt, logfile, defaults_group_suffix,
@@ -1097,7 +1099,7 @@ def cli(database, user, host, port, socket, password, dbname,
         ssl_verify_server_cert, table, csv, warn, execute, myclirc, dsn,
         list_dsn, ssh_user, ssh_host, ssh_port, ssh_password,
         ssh_key_filename, list_ssh_config, ssh_config_path, ssh_config_host,
-        init_command):
+        init_command, charset):
     """A MySQL terminal client with auto-completion and syntax highlighting.
 
     \b
@@ -1222,7 +1224,8 @@ def cli(database, user, host, port, socket, password, dbname,
         ssh_port=ssh_port,
         ssh_password=ssh_password,
         ssh_key_filename=ssh_key_filename,
-        init_command=init_command
+        init_command=init_command,
+        charset=charset
     )
 
     mycli.logger.debug('Launch Params: \n'